fix(datetree): Insert the transclusion link to the end of the entry
The function has failed if there is another transcluded entry
at the insertion position due to read-onlyness.

It was also problematic when there are contents other than transclusion links in
the entry. They would become the content of the transcluded entry.

It is safer to insert the transclusion link after everything in the
entry (except headings).