fix: Add both src lang and mode lang of the context
If the major mode is listed in org-src-lang-modes, add the language name
based on the major mode should be added anyway. This commit adds both languages
(iff they are different from each other) to the language names of the context.

This is useful for preventing from breakage due to an unexpected addition of a
major mode to org-src-lang-modes.